Program conclusion configuration for course groups.
The program conclusion is now a configurable entity. It defines its
name, the graduation title (which can be parameterized with the
${degreeName} variable), if provides alumni, if the final average is
editable in the conclusion process interface and the target registration
state.
New features:
- Added interface to manage program conclusions.
- When creating or editing a course group in a degree curricular plan
the user can select a program conclusion.
- The conclusion process interface provide a list of the available
program conclusions for the current registration, and the user must
select the one to conclude.
- Registry code generator does not depend on cycle type. A default
registry code generator was added with the default implementation.
If necessary for your institution, override the getCode method in
a subclass.
- Requested cycle necessary for document requests (when necessary) is given by
program conclusion instead of cycle type selection.
- Replace calls to degree type unimplemented API to equivalent
methods.
- Remove automatic attribution of senior statute
- Program conclusion can be deleted if not being used in any course
group
- It is possible to link root course groups to program conclusions
- It is possible to add curricular rules to root course groups
- When user overrides the average values use the degree grade scale
for validation purposes
- Users with the academic authorization to repeat the conclusion
process can now revert the last version

Issues: ACDM-686 ACDM-690